Change `eval_fun` parameter to `eval`
You don't evaluate functions anyway - you evaluate calls. So, it was always wrong. But in any case, the thrust is whether or not evaluation occurs, and this now gets at that, even if it comes at the expense of naming an object the same thing as a base R function.